requirements.txt
django~=5.0
djangorestframework~=3.14.0
drf-spectacular~=0.28.0
requests~=2.31.0
pyyaml~=6.0.0
celery[redis]~=5.4.0
redis~=5.2.0
social-auth-app-django~=5.4.0
